refactor(errors): split lib into bridge, pool, proxy, and retry errors
